fix bug

+++ plugins/sticker-toimg.js
@@ -2,7 +2,7 @@ import sharp from 'sharp';
 import { webp2png } from '../lib/webp2mp4.js';
 import fs from 'fs';
 
-const TIMEOUT = 10000; // 10 detik
+const TIMEOUT = 20000; // 20 detik
 
 let handler = async (m, { conn, usedPrefix, command }) => {
   const notStickerMessage = `Reply sticker dengan command *${usedPrefix + command}*`;
@@ -13,41 +13,41 @@ let handler = async (m, { conn, usedPrefix, command }) => {
   const mime = q.mimetype || '';
 
   if (!/image\/webp/.test(mime)) {
-    // Jika sticker tidak dalam format WebP, gunakan fungsi webp2png
     try {
       const name = await conn.getName(m.sender);
       const media = await q.download();
       const out = await webp2png(media).catch(_ => null) || Buffer.alloc(0);
       await conn.sendFile(m.chat, out, 'out.png', 'Request By ' + name, m);
 
-      // Hapus file yang diunduh setelah diunggah
-      await fs.promises.unlink(out || 'out.png' || './tmp/out.png' || `./tmp/${out}`);
+      if (fs.existsSync(out || 'out.png' || './tmp/out.png' || `./tmp/${out}`)) {
+        await fs.promises.unlink(out || 'out.png' || './tmp/out.png' || `./tmp/${out}`);
+      }
     } catch (error) {
       console.error(error);
       m.reply(`Terjadi kesalahan: ${notStickerMessage}`);
     }
   } else {
-    // Jika sticker dalam format WebP, gunakan fungsi sharp untuk konversi
     try {
       const media = await q.download();
       m.reply("Gambar telah diunduh dan OTW dijadikan gambar.");
 
       const decodedBuffer = await sharp(media).toFormat('png')
-        .resize(4096, 4096) // Meningkatkan resolusi menjadi 4K
-        .png({ quality: 100, progressive: true, compressionLevel: 9 }) // Menambahkan opsi untuk kualitas gambar yang sangat tinggi
+        .resize(4096, 4096)
+        .png({ quality: 100, progressive: true, compressionLevel: 9 })
         .toBuffer();
 
       if (decodedBuffer.length > 0) {
         await conn.sendFile(m.chat, decodedBuffer, 'out.png', '*DONE (≧ω≦)ゞ*', m);
 
-        // Hapus file yang diunduh setelah diunggah
-        await fs.promises.unlink('out.png' || './tmp/out.png');
+        if (fs.existsSync('out.png' || './tmp/out.png')) {
+          await fs.promises.unlink('out.png' || './tmp/out.png');
+        }
       } else {
         throw 'Gagal mengonversi stiker menjadi gambar.';
       }
     } catch (error) {
       console.error(error);
-      if (error.message === 'Timeout of 10000ms exceeded') {
+      if (error.message.includes('Timeout')) {
         m.reply('Proses konversi terlalu lama. Silakan coba lagi.');
       } else {
         m.reply(`Terjadi kesalahan: ${notStickerMessage}`);
@@ -58,7 +58,7 @@ let handler = async (m, { conn, usedPrefix, command }) => {
 
 handler.help = ['toimg (reply)'];
 handler.tags = ['sticker'];
-handler.command = ['toimg', 'toimage', 'kegambar', 'toimages','keimg','keimages','togambar'];
+handler.command = ['toimg', 'toimage', 'kegambar', 'toimages', 'keimg', 'keimages', 'togambar'];
 handler.exp = 1500;
 
 export default handler;
